Move and fix tests for validate_existing_memory
