+++ src/vm/runners/builtin_runner.rs
@@ -210,75 +210,4 @@ mod tests {
         let initial_stack = builtin.initial_stack().unwrap();
         assert_eq!(initial_stack.len(), 0);
     }
-    /*
-    #[test]
-    fn validate_existing_memory_for_range_check_within_bounds() {
-        let mut builtin = RangeCheckBuiltinRunner::new(true, bigint!(8), 8);
-        builtin.base = Some(relocatable!(1, 0));
-
-        let mut segments = MemorySegmentManager::new();
-        let mut memory = Memory::new();
-        for _ in 0..3 {
-            segments.add(&mut memory, None);
-        }
-
-        memory
-            .insert(
-                &MaybeRelocatable::from((1, 0)),
-                &MaybeRelocatable::from(bigint!(45)),
-            )
-            .unwrap();
-        let vec = builtin.validate_existing_memory(&memory.data[1]).unwrap();
-        assert_eq!(vec.unwrap()[0], MaybeRelocatable::from((1, 0)));
-    }
-
-    #[test]
-    #[should_panic]
-    fn validate_existing_memory_for_range_check_outside_bounds() {
-        let mut builtin = RangeCheckBuiltinRunner::new(true, bigint!(8), 8);
-        builtin.base = Some(relocatable!(1, 0));
-        let mut memory = Memory::new();
-        memory
-            .insert(
-                &MaybeRelocatable::from((1, 7)),
-                &MaybeRelocatable::from(bigint!(-10)),
-            )
-            .unwrap();
-        builtin.validate_existing_memory(&memory.data[1]).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn validate_existing_memory_for_range_check_relocatable_value() {
-        let mut builtin = RangeCheckBuiltinRunner::new(true, bigint!(8), 8);
-        builtin.base = Some(relocatable!(1, 0));
-        let mut memory = Memory::new();
-        memory
-            .insert(
-                &MaybeRelocatable::from((1, 7)),
-                &MaybeRelocatable::from((1, 4)),
-            )
-            .unwrap();
-        builtin.validate_existing_memory(&memory.data[1]).unwrap();
-    }
-
-    #[test]
-    fn validate_existing_memory_for_range_check_out_of_bounds_diff_segment() {
-        let mut builtin = RangeCheckBuiltinRunner::new(true, bigint!(8), 8);
-        builtin.base = Some(relocatable!(1, 0));
-
-        let mut segments = MemorySegmentManager::new();
-        let mut memory = Memory::new();
-        for _ in 0..3 {
-            segments.add(&mut memory, None);
-        }
-        memory
-            .insert(
-                &MaybeRelocatable::from((2, 0)),
-                &MaybeRelocatable::from(bigint!(-45)),
-            )
-            .unwrap();
-        let vec = builtin.validate_existing_memory(&memory.data[1]);
-        assert_eq!(vec, Ok(None));
-    }*/
 }
